Refactor shared SeaState fusion init and stillness logic
