+++ src/kalman_ou_iii/SeaStateFusionFilter_OU_III.h
@@ -60,6 +60,7 @@
 #include "wave_dir/KalmanWaveDirection.h"
 #include "wave_dir/WaveDirectionDetector.h"
 #include "detrend/AdaptiveWaveDetrender3D.h"
+#include "kalman_common/SeaStateFusionFilterCommon.h"
 
 // Shared constants
 extern const float g_std;
@@ -144,9 +145,10 @@ public:
                     const Eigen::Vector3f& sigma_m)
     {
         mekf_ = std::make_unique<Kalman3D_Wave_OU_III<float>>(sigma_a, sigma_g, sigma_m);
-        enterCold_();      // applies freeze + warmup Racc + disables linear block
-        apply_ou_tune_();
-        mekf_->set_exact_att_bias_Qd(true);
+        seastate::common::finalizeInitialization(
+            mekf_,
+            [this]() { enterCold_(); },
+            [this]() { apply_ou_tune_(); });
     }
 
     void initialize_ext(const Eigen::Vector3f& sigma_a,
@@ -157,9 +159,10 @@ public:
                         float gravity_magnitude)
     {
         mekf_ = std::make_unique<Kalman3D_Wave_OU_III<float>>(sigma_a, sigma_g, sigma_m, Pq0, Pb0, b0, R_S_noise, gravity_magnitude);
-        enterCold_();      // applies freeze + warmup Racc + disables linear block
-        apply_ou_tune_();
-        mekf_->set_exact_att_bias_Qd(true);
+        seastate::common::finalizeInitialization(
+            mekf_,
+            [this]() { enterCold_(); },
+            [this]() { apply_ou_tune_(); });
     }
 
     void initialize_from_acc(const Eigen::Vector3f& acc_body_ned) {
@@ -541,91 +544,8 @@ public:
 private:
 
     // Simple first-order low-pass filter for vertical accel → tracker input
-    struct FreqInputLPF {
-        float state       = 0.0f;
-        float fc_hz       = 1.0f;
-        bool  initialized = false;
-
-        void setCutoff(float fc) {
-            if (std::isfinite(fc) && fc > 0.0f) {
-                fc_hz = fc;
-            }
-        }
-
-        float step(float x, float dt) {
-            const float alpha = std::exp(-2.0f * static_cast<float>(M_PI) * fc_hz * dt);
-
-            if (!initialized) {
-                state       = x;
-                initialized = true;
-                return state;
-            }
-            state = (1.0f - alpha) * x + alpha * state;
-            return state;
-        }
-    };
-
-    // Detect “stillness” from vertical accel and relax frequency when still.
-    struct StillnessAdapter {
-        float energy_ema      = 0.0f;
-        float energy_alpha    = 0.05f;
-        float energy_thresh   = 8e-4f;
-
-        float still_time_sec  = 0.0f;
-        float still_thresh_s  = 2.0f;
-
-        float relax_tau_sec   = 1.0f;
-        float target_freq_hz  = MIN_FREQ_HZ;
-
-        bool  freq_init       = false;
-        float freq_state      = FREQ_GUESS;
-
-        bool  last_is_still   = false;
-
-        void setTargetFreqHz(float f) {
-            if (std::isfinite(f) && f > 0.0f) {
-                target_freq_hz = f;
-            }
-        }
-
-        float step(float a_z_body_proxy_lp, float dt, float freq_in) {
-            if (!(dt > 0.0f) || !std::isfinite(freq_in)) {
-                return freq_in;
-            }
-
-            if (!freq_init || !std::isfinite(freq_state)) {
-                freq_state = freq_in;
-                freq_init  = true;
-            }
-
-            const float a_norm      = a_z_body_proxy_lp / g_std;
-            const float inst_energy = a_norm * a_norm;
-
-            energy_ema = (1.0f - energy_alpha) * energy_ema + energy_alpha * inst_energy;
-            const bool is_still = (energy_ema < energy_thresh);
-            last_is_still = is_still;
-
-            if (is_still) {
-                still_time_sec += dt;
-                if (still_time_sec > 60.0f) still_time_sec = 60.0f;
-
-                if (still_time_sec > still_thresh_s) {
-                    const float relax_alpha = 1.0f - std::exp(-dt / relax_tau_sec);
-                    freq_state += relax_alpha * (target_freq_hz - freq_state);
-                } else {
-                    freq_state = freq_in;
-                }
-            } else {
-                still_time_sec = 0.0f;
-                freq_state     = freq_in;
-            }
-            return freq_state;
-        }
-
-        bool  isStill()       const { return last_is_still; }
-        float getStillTime()  const { return still_time_sec; }
-        float getEnergyEma()  const { return energy_ema; }
-    };
+    using FreqInputLPF = seastate::common::FreqInputLPF;
+    using StillnessAdapter = seastate::common::StillnessAdapter;
 
     void apply_ou_tune_() {
         if (!mekf_) return;
@@ -748,7 +668,7 @@ private:
     void resetTrackingState_() {
         tracker_policy_       = TrackingPolicy{};
         freq_input_lpf_       = FreqInputLPF{};
-        freq_stillness_       = StillnessAdapter{};
+        freq_stillness_       = StillnessAdapter(g_std, min_freq_hz_, FREQ_GUESS);
         freq_input_lpf_.setCutoff(max_freq_hz_);
         freq_stillness_.setTargetFreqHz(min_freq_hz_);
